aprely tabel 2025    KOALA

+++ backasync/src/main/java/com/sous/backasync/launch/interfaces/ModuleInsertingBackAsyncInterface.java
@@ -13,5 +13,5 @@ public interface ModuleInsertingBackAsyncInterface {
 
     Integer getModuleInsert(@NonNull Bundle bundleModuleBack);
     // TODO: 28.01.2025
-    Long getModuleInsert(@NonNull String Таблица, @NonNull ContentValues contentValuesModuleBackAsync);
+    Integer getModuleInsert(@NonNull String Таблица, @NonNull ContentValues contentValuesModuleBackAsync);
 }
